Add files via upload
New task: The program reads in two file names. It copies the first file to the location specified by the second file name.
If the specified file (to be copied) does not exist, then the program should display "File does not exist." and read a file name one more time and only then read the destination file name.